feat: refactoring

+++ src/main/java/com/example/back/workbook/dto/CustomRequest.java
@@ -7,5 +7,5 @@ import java.util.List;
 @Getter
 public class CustomRequest {
     private String title;
-    private List<Integer> selectedQuestionId;
+    private List<Long> selectedQuestionId;
 }
